+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Add `seqs.tree_seq`
+* Add `seqs.tree_seq`, `seqs.empty`
 * Import everything in `clj` as well so one can use `clj.inc` or `clj.first`;
   no need to import sub-modules anymore.